first graph modifications

+++ alg/list/bigString.h
@@ -5,6 +5,6 @@ class bigString : public LinkedList {
   public:
 
     
-    char operator[] (size_t index);
+    char operator[] (int index);
 
 };
